% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addCellColor}
\alias{addCellColor}
\title{Add column colors of an object of ztable}
\usage{
addCellColor(z, rows, cols, color)
}
\arguments{
\item{z}{An object of ztable}

\item{rows}{An integer vector indicating specific rows}

\item{cols}{An integer vector indicating specific columns}

\item{color}{A character vector indicating color}
}
\description{
Add column colors of an object of ztable
}
\examples{
z=ztable(head(iris))
z=addRowColor(z,c(1,3),color="platinum")
z=addColColor(z,2,color="syan")
z=addCellColor(z,cols=c(5,4),rows=5,color="red")
z
}

