% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addcgroup}
\alias{addcgroup}
\title{Add column groups of an object of ztable}
\usage{
addcgroup(z, cgroup, n.cgroup, color = "black", bg = "white", top = FALSE)
}
\arguments{
\item{z}{An object of ztable}

\item{cgroup}{A character vector or matrix indicating names of column group. Default value is NULL}

\item{n.cgroup}{A integer vector or matrix indicating the numbers of columns included in each cgroup
Default value is NULL}

\item{color}{A character vector indicating the font color of each cells.}

\item{bg}{A character vector indicating the background color of each cells.}

\item{top}{Logical. Whether or not cgroup be placed at top.}
}
\description{
Add column groups of an object of ztable
}
