% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addFrontColor}
\alias{addFrontColor}
\title{Add column colors of an object of ztable}
\usage{
addFrontColor(z, rows, cols, color)
}
\arguments{
\item{z}{An object of ztable}

\item{rows}{An integer vector indicating specific rows}

\item{cols}{An integer vector indicating specific columns}

\item{color}{A character vector indicating color}
}
\description{
Add column colors of an object of ztable
}
\examples{
z=ztable(head(iris))
z=addFrontColor(z,rows=2:4,cols=c(2,4,6),color=c("red","green","blue"))
z
}
