% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addrgroup}
\alias{addrgroup}
\title{Add row groups of an object of ztable}
\usage{
addrgroup(z, rgroup, n.rgroup, cspan.rgroup = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{rgroup}{A character vector indicating names of row group. Default value is NULL}

\item{n.rgroup}{A integer vector indicating the numbers of rows included in each rgroup
Default value is NULL}

\item{cspan.rgroup}{An integer indicating the column span of rgroup}
}
\description{
Add row groups of an object of ztable
}
