% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addRowColor}
\alias{addRowColor}
\title{Add row colors of an object of ztable}
\usage{
addRowColor(z, rows = NULL, bg = NULL, color = NULL,
  condition = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{rows}{An integer vector indicating specific rows}

\item{bg}{A character vector indicating background color}

\item{color}{A character vector indicating color}

\item{condition}{Logical expression to select rows}
}
\description{
Add row colors of an object of ztable
}
\examples{
z=ztable(head(iris))
z=addRowColor(z,c(1,3),color="platinum")
z
}
