% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addSigColor}
\alias{addSigColor}
\title{Add row color or cellcolor for rows or cells of p-value less than sigp in a ztable}
\usage{
addSigColor(z, sigp = 0.05, sigcolor = "lightcyan")
}
\arguments{
\item{z}{An object of ztable}

\item{sigp}{A p-value}

\item{sigcolor}{A character indicating color}
}
\description{
Add row color or cellcolor for rows or cells of p-value less than sigp in a ztable
}

