% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addSubColNames}
\alias{addSubColNames}
\title{Add a adjunctive name below colummn name in a ztable}
\usage{
addSubColNames(z, subcolnames)
}
\arguments{
\item{z}{An object of ztable}

\item{subcolnames}{A charactor vector}
}
\description{
Add a adjunctive name below colummn name in a ztable
}
