% Generated by roxygen2 (4.0.2): do not edit by hand
\name{align2html}
\alias{align2html}
\title{Convert the align in Latex format to html format}
\usage{
align2html(align)
}
\arguments{
\item{align}{A character of align in Latex format}
}
\description{
Convert the align in Latex format to html format
}

