% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{align2nd}
\alias{align2nd}
\title{Delete first components of align}
\usage{
align2nd(align)
}
\arguments{
\item{align}{A character for define the align of column in Latex format}
}
\description{
Delete first components of align
}

