% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{alignCount}
\alias{alignCount}
\title{Count the number of align}
\usage{
alignCount(align)
}
\arguments{
\item{align}{A character for define the align of column in Latex format}
}
\description{
Count the number of align
}
