% Generated by roxygen2 (4.0.2): do not edit by hand
\name{caption2minipage}
\alias{caption2minipage}
\title{Convert long caption to minipage}
\usage{
caption2minipage(z, caption)
}
\arguments{
\item{z}{An object of ztable}

\item{caption}{A character vector to convert}
}
\description{
Convert long caption to minipage
}

