% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2flextable.R
\name{cgroup2df}
\alias{cgroup2df}
\title{Convert cgroup of ztable into data.frame}
\usage{
cgroup2df(z)
}
\arguments{
\item{z}{An object of ztable}
}
\value{
A data.frame
}
\description{
Convert cgroup of ztable into data.frame
}
