% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cGroupSpan}
\alias{cGroupSpan}
\title{Count the colspan of each colgroup}
\usage{
cGroupSpan(z)
}
\arguments{
\item{z}{An object of ztable}
}
\value{
A matrix indicating the column span occupied by each colgroup
}
\description{
Count the colspan of each colgroup
}

