% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colGroupCount}
\alias{colGroupCount}
\title{Count the colgroup of an object of ztable}
\usage{
colGroupCount(z)
}
\arguments{
\item{z}{An object of class ztable}
}
\value{
A vector indicating the position of colgroup
}
\description{
Count the colgroup of an object of ztable
}

