% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2flextable.R
\name{color2hex}
\alias{color2hex}
\title{Convert a named color into a hexadecimal color with rgb value}
\usage{
color2hex(color)
}
\arguments{
\item{color}{A named color}
}
\value{
a hexadecimal color
}
\description{
Convert a named color into a hexadecimal color with rgb value
}
\examples{
color2hex("green")
color2hex("red")
}
