% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable.R
\name{data2table}
\alias{data2table}
\title{Convert data to formatted data for table}
\usage{
data2table(z)
}
\arguments{
\item{z}{An object of class "ztable"}
}
\description{
Convert data to formatted data for table
}

