% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{define_colors}
\alias{define_colors}
\title{Define colors}
\usage{
define_colors(mycolors, no = 1)
}
\arguments{
\item{mycolors}{characters vectors of color names}

\item{no}{An integer indicating start number}
}
\description{
Define colors of mycolors
}
