% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Hooks for Namespace Events}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
Functions to be called when loaded, attached, detached or unloaded
}
