% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractAlign}
\alias{extractAlign}
\title{Extract column position information only(without vertical line specifier)}
\usage{
extractAlign(align)
}
\arguments{
\item{align}{A character string indicating align for latex table}
}
\description{
Extract column position information only(without vertical line specifier)
}

