% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNewSpanCol}
\alias{getNewSpanCol}
\title{Calculating new spanCol with spanCol plus space made by column group}
\usage{
getNewSpanCol(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Calculating new spanCol with spanCol plus space made by column group
}

