% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNewSpanRow}
\alias{getNewSpanRow}
\title{Calculating new spanRow with spanRow plus space made by row group}
\usage{
getNewSpanRow(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Calculating new spanRow with spanRow plus space made by row group
}

