% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{getspanRowLength}
\alias{getspanRowLength}
\title{Gets spanRow length}
\usage{
getspanRowLength(z, i, j)
}
\arguments{
\item{z}{An object of ztable}

\item{i}{An integer indicating the row of specific cell}

\item{j}{An integer indicating the column of specific cell}
}
\value{
row count when spanRow starts, 0 when column spans.
}
\description{
Gets spanRow length
}
