% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{gradientColor}
\alias{gradientColor}
\title{Make Sequential colour gradient palette}
\usage{
gradientColor(high = "red", low = "white", mid = NULL, n = 20,
  plot = FALSE)
}
\arguments{
\item{high}{colour for high end of gradient.}

\item{low}{colour for low end of gradient.}

\item{mid}{colour for middle of gradient.}

\item{n}{the number of colors in palette}

\item{plot}{Logical. Whether or not draw plot}
}
\description{
Make Sequential colour gradient palette
}
