% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{hlines}
\alias{hlines}
\title{Add or delete horizontal lines in a ztable}
\usage{
hlines(z, type = NULL, add = NULL, del = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{type}{An integer or one of c("none","all")}

\item{add}{An integer vector indicating rows where the horzontal lines added}

\item{del}{An integer vector indicating rows where the horizontal lines deleted}
}
\description{
Add or delete horizontal lines in a ztable
}

