% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isGroupCol}
\alias{isGroupCol}
\title{Returns whether or not column with position start plus length is group column}
\usage{
isGroupCol(start, length, colCount)
}
\arguments{
\item{start}{An integer indicating start column position}

\item{length}{An integer indicating spanCol length}

\item{colCount}{An integer vector calculating from colGroupCount()}
}
\description{
Returns whether or not column with position start plus length is group column
}

