% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isspanRow}
\alias{isspanRow}
\title{Identify the spanRow status of a cell}
\usage{
isspanRow(z, i, j)
}
\arguments{
\item{z}{An object of ztable}

\item{i}{An integer indicating the row of specific cell}

\item{j}{An integer indicating the column of specific cell}
}
\value{
columns count plus spaces by rgroup when spanRow starts, 0 when row spans,
       minus value when spanRow ends, NULL when no span.
}
\description{
Identify the spanRow status of a cell
}

