% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{make.frontcolor}
\alias{make.frontcolor}
\title{Make a data.frame named "cellcolor" from ztable call}
\usage{
make.frontcolor(z)
}
\arguments{
\item{z}{An object of class "ztable"}
}
\description{
Make a data.frame named "cellcolor" from ztable call
}
