% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.mytable.R
\name{make_align}
\alias{make_align}
\title{Make align for an object of class ztable.mytable}
\usage{
make_align(z)
}
\arguments{
\item{z}{An object of class ztable.mytable}
}
\description{
Make align for an object of class ztable.mytable
}
