% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Add gradient background color to ztable}
\usage{
makeHeatmap(z, palette = "Reds", mycolor = NULL, changeColor = TRUE,
  reverse = FALSE)
}
\arguments{
\item{z}{An object of class ztable}

\item{palette}{Name of color palette}

\item{mycolor}{user defined color vectors}

\item{changeColor}{Logical. Whether of not change font color automatically}

\item{reverse}{If true, reverse the font color}
}
\description{
Add gradient background color to ztable
}
\examples{
require(magrittr)
ztable(head(mtcars)) \%>\% makeHeatmap
require(moonBook)
x=table(acs$smoking,acs$Dx)
ztable(x) \%>\% makeHeatmap
ztable(x) \%>\% makeHeatmap(palette="Blues")
ztable(x) \%>\% makeHeatmap(mycolor=gradientColor(low="yellow",mid="orange",high="red"))
}
