% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{name2rgb}
\alias{name2rgb}
\title{Find rgb value from color name}
\usage{
name2rgb(name)
}
\arguments{
\item{name}{a valid color name}
}
\value{
rgb value
}
\description{
Find rgb value from color name
}

