% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{normalize2}
\alias{normalize2}
\title{Convert numeric vector min to 0, max to maxvalue}
\usage{
normalize2(x, maxvalue = 10)
}
\arguments{
\item{x}{A vector}

\item{maxvalue}{maximal value}
}
\description{
Convert numeric vector min to 0, max to maxvalue
}
