% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelTables.R
\name{parallelTablesLatex}
\alias{parallelTablesLatex}
\title{Place two or more ztables or figures side by side in Latex format}
\usage{
parallelTablesLatex(width, listTables)
}
\arguments{
\item{width}{a numeric vector specifies the width to which the tables or
figures should be scaled}

\item{listTables}{a list consists of object of "ztable" or valid figure name}
}
\description{
Place two or more ztables or figures side by side in HTML format.
The ztable for this purpose should be made by function ztable with tabular="TRUE".
}

