% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable.R
\name{print_ztable}
\alias{print_ztable}
\title{Print an object of class "ztable"}
\usage{
print_ztable(z)
}
\arguments{
\item{z}{An object of class "ztable"}
}
\description{
Print an object of class "ztable"
}

