% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{printHTMLHead}
\alias{printHTMLHead}
\title{Print HTML head if ztable object a has a colgroup}
\usage{
printHTMLHead(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Print HTML head if ztable object a has a colgroup
}

