% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printLatexHead}
\alias{printLatexHead}
\title{Print the head of latex table if the object of ztable has a colgroup}
\usage{
printLatexHead(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Print the head of latex table if the object of ztable has a colgroup
}

