% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printRowGroup}
\alias{printRowGroup}
\title{Print Row Groups in a latex table}
\usage{
printRowGroup(z, i)
}
\arguments{
\item{z}{An object of class ztable}

\item{i}{An integer indicating row}
}
\description{
Print Row Groups in a latex table
}

