% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2flextable.R
\name{roundDf}
\alias{roundDf}
\title{Round the numbers of a data.frame}
\usage{
roundDf(df, digits = 2)
}
\arguments{
\item{df}{A data.frame}

\item{digits}{A vector of integer indicating the number of decimal places}
}
\value{
a rounded data.frame
}
\description{
Round the numbers of a data.frame
}
