% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spanColWidth}
\alias{spanColWidth}
\title{Calculate the spanColWidth when spanCol start}
\usage{
spanColWidth(z, i, j)
}
\arguments{
\item{z}{An object of ztable}

\item{i}{An integer indicating the row of specific cell}

\item{j}{An integer indicating the column of specific cell}
}
\value{
column count when spanCol start
}
\description{
Calculate the spanColWidth when spanCol start
}

