% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spanRow}
\alias{spanRow}
\title{Merging data cells of ztable object in rows}
\usage{
spanRow(z, col, from, to, color = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{col}{An integer indicating the column of merging data cell}

\item{from}{An integer indicating start row of merging data cell}

\item{to}{An integer indicating end row of merging data cell}

\item{color}{An optional character indicating the background color of merging cell}
}
\description{
Merging data cells of ztable object in rows
}

