% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable.R
\name{tableLength}
\alias{tableLength}
\title{Convert data to formatted data for table}
\usage{
tableLength(z)
}
\arguments{
\item{z}{An object of class "ztable"}
}
\description{
Convert data to formatted data for table
}

