% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.mytable.R
\name{trim.ztable}
\alias{trim.ztable}
\title{Make align and edit p value column for an object of class ztable.mytable}
\usage{
trim.ztable(z)
}
\arguments{
\item{z}{An object of class ztable.mytable}
}
\description{
Make align and edit p value column for an object of class ztable.mytable
}
