% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable.R
\name{validColor2}
\alias{validColor2}
\title{Find valid color name}
\usage{
validColor2(a)
}
\arguments{
\item{a}{An integer or a character}
}
\value{
a valid Latex color name
}
\description{
Find valid color name
}

