% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vline2align}
\alias{vline2align}
\title{Make a latex "align" from a string and vertical line specifier}
\usage{
vline2align(align, vlines)
}
\arguments{
\item{align}{A character string indicating align of latex table}

\item{vlines}{An integer vector indicating vertical line position}
}
\description{
Make a latex "align" from a string and vertical line specifier
}

