% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vlines}
\alias{vlines}
\title{Add or delete vertical lines in a ztable}
\usage{
vlines(z, type = NULL, add = NULL, del = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{type}{An integer or one of c("none","all","group","nongroup")}

\item{add}{An integer vector indicating columns where the width of vertical lines added}

\item{del}{An integer vector indicating columns where the width of vertical lines subtracted}
}
\description{
Add or delete vertical lines in a ztable
}

