% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcolors.R
\docType{data}
\name{zcolors}
\alias{zcolors}
\title{Definition of Latex Color}
\format{A data frame with 749 rows and 3 variables:
\describe{
  \item{name}{Color name}
  \item{rgb}{Hex triplet of color}
  \item{definition}{Latex command of color definition}
}}
\usage{
zcolors
}
\description{
A dataset cotaining the name of color and Hex-triplets and latex definition
}
\details{
To use this color definition, a latex package "color" shoule bd included in
your preamble.
}
\keyword{datasets}
