% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ztable.cbind.mytable}
\alias{ztable.cbind.mytable}
\title{Make ztable from object cbind.mytable}
\usage{
\method{ztable}{cbind.mytable}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An object of cbind.mytable}

\item{digits}{Numeric vector of length equal to one (in which case it will be
replicated as necessary) or to the number of columns of the resulting table}

\item{...}{arguments to be passed to \code{\link{ztable_sub}}}
}
\description{
Make ztable from object cbind.mytable
}
\examples{
require(moonBook)
res=mytable(sex+DM~.,data=acs)
z=ztable(res)
z
}

