% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable.mytable.R
\name{ztable.mytable}
\alias{ztable.mytable}
\title{Make ztable from object mytable}
\usage{
\method{ztable}{mytable}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An object of mytable}

\item{digits}{Numeric vector of length equal to one (in which case it will be
replicated as necessary) or to the number of columns of the resulting table}

\item{...}{arguments to be passed to \code{\link{ztable_sub}}}
}
\description{
Make ztable from object mytable
}
\examples{
require(moonBook)
res=mytable(sex~.,data=acs)
z=ztable(res)
z
}

