% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2flextable.R
\name{ztable2flextable}
\alias{ztable2flextable}
\title{Convert an object of ztable into an object of flextable}
\usage{
ztable2flextable(z)
}
\arguments{
\item{z}{An object of class ztable}
}
\value{
An object of class flextable
}
\description{
Convert an object of ztable into an object of flextable
}
\examples{
z=ztable(head(mtcars))
ztable2flextable(z)
}
