% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{ztable2html}
\alias{ztable2html}
\title{Print an object of class "ztable" to html table}
\usage{
ztable2html(z, xdata)
}
\arguments{
\item{z}{An object of class "ztable"}

\item{xdata}{A formatted data.frame}
}
\description{
Print an object of class "ztable" to html table
}

