% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{ztable2viewer}
\alias{ztable2viewer}
\title{Print an object of ztable via rstudio::viewer}
\usage{
ztable2viewer(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Print an object of ztable via rstudio::viewer
}

