% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addcgroup}
\alias{addcgroup}
\title{Add column groups of an object of ztable}
\usage{
addcgroup(z, cgroup, n.cgroup, cgroupcolor = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{cgroup}{A character vector or matrix indicating names of column group. Default value is NULL}

\item{n.cgroup}{A integer vector or matrix indicating the numbers of columns included in each cgroup
Default value is NULL}

\item{cgroupcolor}{A character vector or matrix indicating the background colors of each cells.}
}
\description{
Add column groups of an object of ztable
}
