% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addColColor}
\alias{addColColor}
\title{Add column colors of an object of ztable}
\usage{
addColColor(z, cols, color)
}
\arguments{
\item{z}{An object of ztable}

\item{cols}{An integer vector indicating specific columns}

\item{color}{A character vector indicating color}
}
\description{
Add column colors of an object of ztable
}
\examples{
z=ztable(head(iris))
z=addColColor(z,c(1,3),color="platinum")
z
}

