% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable2.R
\name{addSigColor}
\alias{addSigColor}
\title{Add row color or cellcolor for rows or cells of p-value less than sigp in a ztable}
\usage{
addSigColor(z, level = 0.05, bg = "lightcyan", color = "black")
}
\arguments{
\item{z}{An object of ztable}

\item{level}{A p-value}

\item{bg}{A character indicating background color}

\item{color}{A character indicating color}
}
\description{
Add row color or cellcolor for rows or cells of p-value less than sigp in a ztable
}
