% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addSubColNames}
\alias{addSubColNames}
\title{Add a adjunctive name below colummn name in a ztable}
\usage{
addSubColNames(z, subcolnames)
}
\arguments{
\item{z}{An object of ztable}

\item{subcolnames}{A charactor vector}
}
\description{
Add a adjunctive name below colummn name in a ztable
}

