% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{align2lines}
\alias{align2lines}
\title{count the vertical column lines from align of Latex format}
\usage{
align2lines(align)
}
\arguments{
\item{align}{A string of align Latex format}
}
\value{
a numeric vector consists of vertical lines of each column
}
\description{
count the vertical column lines from align of Latex format
}

