% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alignCheck}
\alias{alignCheck}
\title{Check the validity of align}
\usage{
alignCheck(align, ncount, addrow)
}
\arguments{
\item{align}{A character for define the align of column in Latex format}

\item{ncount}{An integer equals of ncol function}

\item{addrow}{An integer}
}
\description{
Check the validity of align
}

