% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{caption2minipage}
\alias{caption2minipage}
\title{Convert long caption to minipage}
\usage{
caption2minipage(z, caption)
}
\arguments{
\item{z}{An object of ztable}

\item{caption}{A character vector to convert}
}
\description{
Convert long caption to minipage
}
