% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{colGroupCount}
\alias{colGroupCount}
\title{Count the colgroup of an object of ztable}
\usage{
colGroupCount(z)
}
\arguments{
\item{z}{An object of class ztable}
}
\value{
A vector indicating the position of colgroup
}
\description{
Count the colgroup of an object of ztable
}

