% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable.R
\name{define_colors}
\alias{define_colors}
\title{Define colors}
\usage{
define_colors(mycolors)
}
\arguments{
\item{mycolors}{chracters vectors of color names}
}
\description{
Define colors of mycolors
}

