% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNewAlign}
\alias{getNewAlign}
\title{Make a charater string indicating the alignment of components of table.}
\usage{
getNewAlign(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Make a charater string indicating the alignment of components of table.
}

