% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ztable.html.R
\name{getNewAlign}
\alias{getNewAlign}
\title{Make a character string indicating the alignment of components of table.}
\usage{
getNewAlign(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Make a character string indicating the alignment of components of table.
}
