% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{getNewSpanRow}
\alias{getNewSpanRow}
\title{Calculating new spanRow with spanRow plus space made by row group}
\usage{
getNewSpanRow(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Calculating new spanRow with spanRow plus space made by row group
}

