% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{getspanRowData}
\alias{getspanRowData}
\title{Gets the spanRaw start column}
\usage{
getspanRowData(z, i, j)
}
\arguments{
\item{z}{An object of ztable}

\item{i}{An integer indicating the row of specific cell}

\item{j}{An integer indicating the column of specific cell}
}
\value{
An integer indicating column where spanRaw start. This function is for latex
       multirow
}
\description{
Gets the spanRaw start column
}

