% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{isspanCol}
\alias{isspanCol}
\title{Idetify the spanCol status of a cell}
\usage{
isspanCol(z, i, j)
}
\arguments{
\item{z}{An object of ztable}

\item{i}{An integer indicating the row of specific cell}

\item{j}{An integer indicating the column of specific cell}
}
\value{
column plus space count when spanCol starts, 0 when column spans,
       minus value when spanCol ends, NULL when no span.
}
\description{
Idetify the spanCol status of a cell
}

