% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{make.frontcolor}
\alias{make.frontcolor}
\title{Make a data.frame named "cellcolor" from ztable call}
\usage{
make.frontcolor(x, color = "black")
}
\arguments{
\item{x}{A data.frame}

\item{color}{A character string}
}
\description{
Make a data.frame named "cellcolor" from ztable call
}
