% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable.R
\name{print.ztable}
\alias{print.ztable}
\title{Print an object of class "ztable"}
\usage{
\method{print}{ztable}(x, ...)
}
\arguments{
\item{x}{An object of class "ztable"}

\item{...}{further argument passed to other function}
}
\description{
Print an object of class "ztable"
}

