% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{printLatexHead}
\alias{printLatexHead}
\title{Print the head of latex table if the object of ztable has a colgroup}
\usage{
printLatexHead(z)
}
\arguments{
\item{z}{An object of ztable}
}
\description{
Print the head of latex table if the object of ztable has a colgroup
}

