% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable.R
\name{repColor}
\alias{repColor}
\title{Make vector x from vector color}
\usage{
repColor(x, color)
}
\arguments{
\item{x}{A destination vector}

\item{color}{A charactor vector consists of color names}
}
\description{
Internal function of make.cell.color
}

