% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztable2.R
\name{spanCol}
\alias{spanCol}
\title{Merging data cells of ztable object in columns}
\usage{
spanCol(z, row, from, to, color = NULL)
}
\arguments{
\item{z}{An object of ztable}

\item{row}{An integer indicating the row of merging data cell}

\item{from}{An integer indicating start column of merging data cell}

\item{to}{An integer indicating end column of merging data cell}

\item{color}{An optional character indicating the background color of merging cell}
}
\description{
Merging data cells of ztable object in columns
}

