% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.mytable.R
\name{totalLeft}
\alias{totalLeft}
\title{Arrage total column to the left}
\usage{
totalLeft(z)
}
\arguments{
\item{z}{An object of class ztable.mytable or ztable.cbind.mytable}
}
\description{
Arrage total column to the left
}
\examples{
require(moonBook)
require(ztable)
require(magrittr)
mytable(sex~.,data=acs,show.total=TRUE) \%>\% ztable \%>\% totalLeft
mytable(sex+Dx~.,data=acs,show.total=TRUE) \%>\% ztable \%>\% totalLeft
}
