% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztable.R
\name{validColor}
\alias{validColor}
\title{Find valid color name}
\usage{
validColor(a, mycolor)
}
\arguments{
\item{a}{An integer or a character}

\item{mycolor}{predefined color names}
}
\value{
a valid Latex color name
}
\description{
Find valid color name
}
