% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R, R/ztable.R, R/ztable1.R
\name{ztable.table}
\alias{ztable.table}
\alias{ztable}
\alias{ztable.default}
\alias{ztable.data.frame}
\alias{ztable.matrix}
\alias{ztable.lm}
\alias{ztable.fitdistr}
\alias{ztable.nls}
\alias{ztable.aov}
\alias{ztable.anova}
\alias{ztable.glm}
\alias{ztable.coxph}
\alias{ztable.prcomp}
\alias{ztable.summary.prcomp}
\title{Exporting a R object to an object of class "ztable"}
\usage{
\method{ztable}{table}(x, digits = NULL, ...)

ztable(x, digits = NULL, ...)

\method{ztable}{default}(x, digits = NULL, ...)

\method{ztable}{data.frame}(x, digits = NULL, ...)

\method{ztable}{matrix}(x, digits = NULL, ...)

\method{ztable}{lm}(x, digits = NULL, ...)

\method{ztable}{fitdistr}(x, digits = NULL, ...)

\method{ztable}{nls}(x, digits = NULL, ...)

\method{ztable}{aov}(x, digits = NULL, ...)

\method{ztable}{anova}(x, digits = NULL, ...)

\method{ztable}{glm}(x, digits = NULL, ...)

\method{ztable}{coxph}(x, digits = NULL, ...)

\method{ztable}{prcomp}(x, digits = NULL, ...)

\method{ztable}{summary.prcomp}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An R object, mainly data.frame}

\item{digits}{Numeric vector of length equal to one (in which case it will be
replicated as necessary) or to the number of columns of the resulting table}

\item{...}{arguments to be passed to \code{\link{ztable_sub}}}
}
\description{
Exporting a R object to an object of class "ztable"
}
\section{Methods (by class)}{
\itemize{
\item \code{table}: Makes a ztable for class table

\item \code{default}: Default method of ztable

\item \code{data.frame}: Makes a ztable for class 'data.frame'

\item \code{matrix}: Makes a ztable for class matrix

\item \code{lm}: Makes a ztable for class 'lm'

\item \code{fitdistr}: Makes a ztable for class 'fitdistr'

\item \code{nls}: Makes a ztable for class 'nls'

\item \code{aov}: Makes a ztable for class 'aov'

\item \code{anova}: Makes a ztable for class 'anova'

\item \code{glm}: Makes a ztable for class 'glm'

\item \code{coxph}: Makes a ztable for class 'coxph'

\item \code{prcomp}: Makes a ztable for class 'prcomp'

\item \code{summary.prcomp}: Makes a ztable for class 'summary.prcomp'
}}

