% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ztable}
\alias{ztable}
\alias{ztable.anova}
\alias{ztable.aov}
\alias{ztable.coxph}
\alias{ztable.data.frame}
\alias{ztable.default}
\alias{ztable.glm}
\alias{ztable.lm}
\alias{ztable.matrix}
\alias{ztable.prcomp}
\alias{ztable.summary.prcomp}
\title{Exporting a R object to an object of class "ztable"}
\usage{
ztable(x, digits = NULL, ...)

\method{ztable}{default}(x, digits = NULL, ...)

\method{ztable}{data.frame}(x, digits = NULL, ...)

\method{ztable}{matrix}(x, digits = NULL, ...)

\method{ztable}{lm}(x, digits = NULL, ...)

\method{ztable}{aov}(x, digits = NULL, ...)

\method{ztable}{anova}(x, digits = NULL, ...)

\method{ztable}{glm}(x, digits = NULL, ...)

\method{ztable}{coxph}(x, digits = NULL, ...)

\method{ztable}{prcomp}(x, digits = NULL, ...)

\method{ztable}{summary.prcomp}(x, digits = NULL, ...)
}
\arguments{
\item{x}{An R object, mainly data.frame}

\item{digits}{Numeric vector of length equal to one (in which case it will be
replicated as necessary) or to the number of columns of the resulting table}

\item{...}{arguments to be passed to \code{\link{ztable_sub}}}
}
\description{
Exporting a R object to an object of class "ztable"
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{data.frame}: 

\item \code{matrix}: 

\item \code{lm}: 

\item \code{aov}: 

\item \code{anova}: 

\item \code{glm}: 

\item \code{coxph}: 

\item \code{prcomp}: 

\item \code{summary.prcomp}: 
}}

